requirements.txt
frozendict
requests
inflection
urllib3
python-dateutil
pydantic
fastapi
python-jose[cryptography]
passlib[bcrypt]
ujson
jinja2
aiofiles
fastapi-crudrouter
motor
git+https://github.com/archetipo/python-formio-data.git@imp_resource_ext
pytz
json2html
pandas
pymongo
ODMantic
python-slugify
python-multipart
json_logic